move: 편지읽기 팝업 관련 파일 폴더 이동
