+++ frontend/src/components/MainPage/readLetter/CloseButton.jsx
