+++ frontend/src/components/MainPage/readLetter/LetterPopLabel.jsx
@@ -1,0 +1,14 @@
+import LetterPopTextArea from "@components/MainPage/readLetter/LetterPopTextArea";
+
+const LetterPopLabel = ({ label, text }) => {
+  return (
+    <div className="flex items-center my-[5%]">
+      <div className="w-[35%]">{label}</div>
+      <div className="w-[65%]">
+        <LetterPopTextArea text={text} />
+      </div>
+    </div>
+  );
+};
+
+export default LetterPopLabel;
